+++ CHANGELOG.md
@@ -2,3 +2,4 @@ v0.1.0
 - n-gram tokenizer
 - LMDB dataset and dataloader
 - Pytorch Lightining trainer
+- preprocessed Ratings Dataset